Fixed incorrect return type

+++ src/libs/ino_libs/ino_libs.h
@@ -16,7 +16,7 @@ void pinMode(volatile uint8_t *PORT, uint8_t pin, uint8_t mode);
 void digitalWrite(volatile uint8_t *PORT, uint8_t pin, uint8_t value);
 void analogWrite(volatile uint8_t *PORT, uint8_t pin, uint8_t value);
 uint8_t digitalRead(volatile uint8_t *PORT, uint8_t pin);
-uint8_t analogRead(volatile uint8_t *PORT, uint8_t pin);
+uint16_t analogRead(volatile uint8_t *PORT, uint8_t pin);
 uint32_t pulseIn(volatile uint8_t *PORT, uint8_t pin, uint8_t state, uint32_t timeout, void (*sendPulse)());
 
 #endif
